elli (elli v3.3.0)

Elli acceptor manager

This gen_server owns the listen socket and manages the processes accepting on that socket. When a process waiting for accept gets a request, it notifies this gen_server so we can start up another acceptor.

Types

body/0

-type body() :: binary() | iolist().

A binary or iolist.

header/0

-type header() :: {Key :: binary(), Value :: binary() | string()}.

headers/0

-type headers() :: [header()].

http_method/0

-type http_method() :: 'OPTIONS' | 'GET' | 'HEAD' | 'POST' | 'PUT' | 'DELETE' | 'TRACE' | binary().

An uppercase atom representing a known HTTP verb or a binary for other verbs.

req/0

-type req() ::
          #req{method :: elli:http_method(),
               scheme :: undefined | binary(),
               host :: undefined | binary(),
               port :: undefined | 1..65535,
               path :: [binary()],
               args :: [{binary(), any()}],
               raw_path :: binary(),
               version :: elli_http:version(),
               headers :: elli:headers(),
               original_headers :: elli:headers(),
               body :: elli:body(),
               pid :: pid(),
               socket :: undefined | elli_tcp:socket(),
               callback :: elli_handler:callback()}.

response_code/0

-type response_code() :: 100..999.

state/0

-opaque state()

Internal state.

Functions

get_acceptors(S)

-spec get_acceptors(atom()) -> {reply, {ok, [ets:tid()]}, state()}.

get_open_reqs(S)

-spec get_open_reqs(S :: atom()) -> {reply, {ok, non_neg_integer()}, state()}.

get_open_reqs(S, Timeout)

-spec get_open_reqs(S :: atom(), Timeout :: non_neg_integer()) -> Reply
                       when Reply :: {reply, {ok, non_neg_integer()}, state()}.

set_callback(S, Callback, CallbackArgs)

-spec set_callback(S, Callback, CallbackArgs) -> Reply
                      when
                          S :: atom(),
                          Callback :: elli_handler:callback_mod(),
                          CallbackArgs :: elli_handler:callback_args(),
                          Reply :: {reply, ok, state()}.

start_link()

-spec start_link() -> Result
                    when
                        Result :: {ok, Pid} | ignore | {error, Error},
                        Pid :: pid(),
                        Error :: {already_started, Pid} | term().

Create an Elli server process as part of a supervision tree, using the default configuration. The same as start_link({callback, elli_example_callback}, {callback_args, []}).

start_link(Opts)

-spec start_link(Opts) -> Result
                    when
                        Opts :: [{_, _}],
                        Result :: {ok, Pid} | ignore | {error, Error},
                        Pid :: pid(),
                        Error :: {already_started, Pid} | term().

stop(Server)

-spec stop(Server :: atom()) -> ok.

Stop Server.