elli_http (elli v3.3.0)

Elli HTTP request implementation

An elli_http process blocks in elli_tcp:accept/3 until a client connects. It then handles requests on that connection until it's closed either by the client timing out or explicitly by the user.

Types

version/0

-type version() :: {0, 9} | {1, 0} | {1, 1}.

HTTP version as a tuple, i.e. {0, 9} | {1, 0} | {1, 1}.

Functions

accept(Server, ListenSocket, Options, Callback)

-spec accept(Server, ListenSocket, Options, Callback) -> ok
                when
                    Server :: pid(),
                    ListenSocket :: elli_tcp:socket(),
                    Options :: proplists:proplist(),
                    Callback :: elli_handler:callback().

Accept on the socket until a client connects. Handle the request, then loop if we're using keep alive or chunked transfer. If elli_tcp:accept/3 doesn't return a socket within a configurable timeout, loop to allow code upgrades of this module.

chunk_loop(Socket)

close_or_keepalive(Req, Headers)

-spec close_or_keepalive(Req, Headers) -> KeepaliveOpt
                            when
                                Req :: elli:req(),
                                Headers :: elli:headers(),
                                KeepaliveOpt :: close | keep_alive.

Return the preferred session handling setting to close or keep the current session alive based on the presence of a header or the standard default based on the version of HTTP of the request.

handle_request(Socket, PrevBin, Options, Callback)

-spec handle_request(Socket, PrevBin, Options, Callback) -> ConnToken
                        when
                            Socket :: elli_tcp:socket(),
                            PrevBin :: binary(),
                            Options :: proplists:proplist(),
                            Callback :: elli_handler:callback(),
                            ConnToken :: {keep_alive | close, binary()}.

Handle a HTTP request that will possibly come on the socket. Returns the appropriate connection token and any buffer containing (parts of) the next request.

keepalive_loop(Socket, Options, Callback)

Handle multiple requests on the same connection, i.e. "keep alive".

keepalive_loop(Socket, NumRequests, Buffer, Options, Callback)

mk_req(Method, PathTuple, Headers, Headers, Body, V, Socket, Callback)

-spec mk_req(Method, PathTuple, Headers, Headers, Body, V, Socket, Callback) -> Req
                when
                    Method :: elli:http_method(),
                    PathTuple :: {PathType :: atom(), RawPath :: binary()},
                    Headers :: elli:headers(),
                    Body :: elli:body(),
                    V :: version(),
                    Socket :: elli_tcp:socket() | undefined,
                    Callback :: elli_handler:callback(),
                    Req :: elli:req().

mk_req(Method, Scheme, Host, Port, PathTuple, Headers, ParsedHeaders, Body, V, Socket, Callback)

parse_path(_)

send_file(Req, Code, Headers, Filename, Range)

-spec send_file(Req, Code, Headers, Filename, Range) -> ok
                   when
                       Req :: elli:req(),
                       Code :: elli:response_code(),
                       Headers :: elli:headers(),
                       Filename :: file:filename(),
                       Range :: elli_util:range().

Send a HTTP response to the client where the body is the contents of the given file. Assumes correctly set response code and headers.

send_response(Req, Code, Headers, UserBody)

Generate a HTTP response and send it to the client.

split_args(Qs)

-spec split_args(binary()) -> [{binary(), binary() | true}].

Split the URL arguments into a proplist. Lifted from cowboy_http:x_www_form_urlencoded/2.

start_link(Server, ListenSocket, Options, Callback)

-spec start_link(Server, ListenSocket, Options, Callback) -> pid()
                    when
                        Server :: pid(),
                        ListenSocket :: elli_tcp:socket(),
                        Options :: proplists:proplist(),
                        Callback :: elli_handler:callback().