elli_request (elli v3.3.0)

Types

http_range/0

-type http_range() ::
          {First :: non_neg_integer(), Last :: non_neg_integer()} |
          {offset, Offset :: non_neg_integer()} |
          {suffix, Length :: pos_integer()}.

Functions

async_send_chunk(Ref, Data)

Send a chunk asynchronously.

body(Req)

Return the body.

body_qs(Req)

Parse application/x-www-form-urlencoded body into a proplist.

chunk_ref(Req)

Return a reference that can be used to send chunks to the client. If the protocol does not support it, return {error, not_supported}.

close_chunk(Ref)

Explicitly close the chunked connection. Return {error, closed} if the client already closed the connection. The same as send_chunk(Ref, close)

get_arg(Key, Req)

get_arg(Key, Req, Default)

get_arg_decoded(Key, Req)

get_arg_decoded(Key, Req, Default)

get_args(Req)

-spec get_args(elli:req()) -> QueryArgs :: proplists:proplist().

Return a proplist of keys and values of the original query string. Both keys and values in the returned proplists will be binaries or the atom true in case no value was supplied for the query value.

get_args_decoded(Req)

get_header(Key, Req)

get_header(Key, Req, Default)

get_range(Req)

-spec get_range(elli:req()) -> [http_range()] | parse_error.

Parse the Range header from the request. The result is either a [http_range()] or the atom parse_error. Use elli_util:normalize_range/2 to get a validated, normalized range.

headers(Req)

Return the headers that have had string:casefold/1 run on each key.

host(Req)

Return the host.

is_request(Req)

method(Req)

Return the method.

original_headers(Req)

Return the original headers.

path(Req)

Return path split into binary parts.

peer(Req)

port(Req)

Return the port.

post_arg(Key, Req)

post_arg(Key, Req, Default)

post_arg_decoded(Key, Req)

post_arg_decoded(Key, Req, Default)

post_args(Req)

post_args_decoded(Req)

query_str(Req)

-spec query_str(elli:req()) -> QueryStr :: binary().

Calculate the query string associated with a given Request as a binary.

raw_path(Req)

Return the raw_path, i.e. not split or parsed for query params.

scheme(Req)

Return the scheme.

send_chunk(Ref, Data)

Send a chunk synchronously. If the referenced process is dead, return early with {error, closed}, instead of timing out.

to_proplist(Req)

Serialize the Request record to a proplist. Useful for logging.

uri_decode(Bin)